Suppress configuration input tracking in vintage mode too
Input Tracking affects not only what goes into the CC fingerprint but
also what code is allowed to do. In particular, code that contributes
to inputs must not start external processes through Java API. With the
newly introduced vintage mode deprecation of such behavior it is now
important to maintain the proper input tracking suppression state there,
as the deprecation-emitting infra relies on the flag.

This commit hoists the toggle point out of CC-only code into common.
